Add missing policycoreutils-python package (#86)
* Add missing policycoreutils-python package

In addition:

- Include additional platforms and suites for testing
- Run `apt_update` in `selinux_install` on Debian-based systems

Signed-off-by: Lance Albertson <lance@osuosl.org>

* Fix SELinux enablement on Ubuntu 18.04

For whatever reason Ubuntu 18.04 doesn't properly relabel after the first
reboot. To fix this, we need to check to see if some critical binaries (i.e.
sshd) are properly labeled, and if not relabel and reboot again. This second
reboot seems to properly fix the issues with this release of Ubuntu.

In addition, it seems use of the selinux-activate binary is also breaking some
of this, so for now just use touch /.autorelabel. Also, the kitchenConverge
SELinux module needed some adjustments to work properly on this platform.

Let's also reduce the wait for retry to 180 seconds to speed up CI tests.

Signed-off-by: Lance Albertson <lance@osuosl.org>

Signed-off-by: Lance Albertson <lance@osuosl.org>